New pc.net.OAuth() and pc.net.Http() classes
Refactored net_http so that there is now a pc.net.Http class and a default instance of this is created at pc.net.http. So no compatability issues.

Added pc.net.OAuth which sub-classes pc.net.Http and can be used to access resources that require OAuth 2.0 authentication. Supports the same get/post/put/delete methods as Http, but adds a refreshAccessToken() method which is used to get a new valid access token. Uses client-side cookies for authentication.
A 401 response will automatically trigger an refresh of the access token and the request will be resent.