Resource Requests are not typed and cache hits only occur if type matches
You can define an optional Type value on a ResourceRequest, if you do, values in the cache will be checked against this type, and if they don't match the cache value is not sued.
This prevents ImageRequests and TextureRequests getting confused in the cache. Though it means they will overwrite each other in the cache...
Fixes skyboxes

--HG--
branch : feature/json-loader