Add pc.Entity#addComponent() new pc.Entity() now requires application as argument
With this change you can now call addComponent on an Entity object. This is less verbose and more understandable. (Adding a component to an entity, rather than and an entity to a component system.)

This should be a backwards compatible change as Entity will (temporarily) retrieve the application using the standard canvas id which is used in online play canvas games.
There is a deprecation warning that all new Entities should now be passed the application.