Add optional name-parameter to Entity/Graph constructor (#845)
This PR saves lots of coding time by allowing developers to pass the entity name directly in the constructor. This not only minimizes tedious code but also makes its easier for the many that are familiar with Unity and/or porting code from Unity, see [`GameObject` constructor](https://docs.unity3d.com/ScriptReference/GameObject-ctor.html).

Please note that this still supports the old **undocumented** API of passing the `pc.Application` to the constructor. The PR is implemented to correctly handle the all use cases. In addition, the PR also documents all parameters (not just the new optional name) for both the `Entity` and `GraphNode` constructors and all exempels have also been updated to the comply with the documented API.

Summary of PR:
- Completely backwards-compatible.
- Full API documentation.
- Well tested!
- New optional APIs:
- - `new pc.Entity(name, app)` (old `new pc.Entity(app)` is still supported!)
- - `new pc.GraphNode(name)`


Please consider this important PR for us. Cheers