Fix SoundInstance.sound set under !hasAudioContext.
In the hasAudioContext() branch, SoundInstance.sound = whatever will call _createSource, which sets SoundInstance.source.

If WebAudio isn't around and hasAudioContext() is false, we still need to call _createSource to do this, lest subsequent "play" and other methods exit early.

This was observed in a project where we'd unchecked "preload" on a number of large voiceover and music audio assets, and then ran under Cocoon's Developer App in Canvas+ mode, which does not support WebAudio (and hence is hasAudioContext=false).

When the sound load completed and SoundSlot._onAssetLoad triggered the onLoad handler (added earlier in SoundSlot.play), it would successfully set instance.sound but not instance.source, so the non-WebAudio instance.play() would exit immediately.

It's possible the _createSource calls from the WebAudio side in play and resume should be migrated over to the non-WebAudio side as well.