Prevent initialize() overwriting textures to null if they are waiting on an asset to be loaded
initialize() is called on a material immediately after the placeholder textures are assigned. The source data currently replaces asset id with the actual texture and initialize would clear to null if the data didn't contain the texture. Instead we now also check if it is an asset id, in which case we can wait until the asset is loaded and leave the placeholder texture.

This prevents additional shader compilation calls (because the number of textures on the material changes).