Replace placeholder values for torus primitive area according to suggestions from @guycalledfrank
"I guess _area.uv is 2/3 because it was unwrapped like that, taking 2/3 of UV square (2 rows with 3 faces, one empty row). _area xyz is separated because we allow non-uniform scale. and xyz are probably each = 2 because we have 2 faces with area=1 on each side (left+right, top+bottom, front+back). so _area.xyz is the sum of projected area for each axis (including backfaces).

For a capsule, I'm not sure it looks right. I think it should be divided by 2. And even so, for side projections it won't be exactly right."

see https://github.com/playcanvas/engine/pull/1842#issuecomment-579311346