Fixed First Person Jump
When the player jumps and falls, it goes up normally and falls with a normal behaviour in the first half of the fall but as it approaches the ground it slows down and you can can see it falls really slowly until it touches the ground. This happens every time you jump.

This commit fixes that.