Fixed assets.find to match APIs and return null if asset is not found
The docs say that assets.find should return null if the asset is not found and the original code implies that null should be returned.

This is a potentially breaking change. I don't think undefined should be a valid return type in general but very wary about merging this as developers could be relying on the current behaviour of returning undefined