Added same color and same sprite check in Button component
Fixes #3814

Breaks cyclic call when there are multiple button components referring to the same image component

Between this PR and #3761, the check on whether the same sprite/color is being set on the element component has moved from the element component to the button component. The button component is now responsible for breaking the cyclic call.

Thank you @Sarletor for the fix on the color in PR #3814