Rewrote the way we linearize depth.
Actually, orthographic projection already stores the depth linearly, so all we need to do is to scale it back so it's between near and far plane.

As for perspective projection, since the depth is already in normalized space [0, 1], there should be no need to move it back to NDC coordinates since we don't want the depth in -1, 1 when we output.

One unfortunate effect is that we are more or less forced to branch on a uniform since the shader has no awareness of the type of projection being used.