Fresnel refactor, f0 tint and specularity factor.
Added support for KHR_materials_specular, which provides a specular tint (same as specularMap) as well as a specularity factor. The specularity factor isn't a substitute for metalness however, but simply a specular reflection intensity, which is applied to both fresnel and general specularity.

Decoupled the fresnel component from the specularity. This is because fresnel is dependent on each light, and in the case of reflections dependent on the surface normal. Therefore, we should calculate the fresnel factor once per each light source. Currently though, the fresnel factor only applies to reflections, but that's the same behaviour from before, and this should be addressed in a separate ticket.

Metalness modulation (linear interpolation between f0 and albedo) is done in the backend as a separate shader chunk, and metalness is passed to the backend with dMetalness.

Since the specular color, f0, is dependent on the fresnel effect, and the fresnel effect is dependent on a per-light basis, the specular color can't just be applied on top of everything when combining the light together. Instead, the f0 color has to be an argument to each specularity calculation for it to be accurate.

We can also remove the dSpecularityNoFres from the area light calculation as the fresnel function is no longer combined with the f0 color.