Calculate fresnel for directional lights
This is necessary to not get too much light when using directional lights.
Calculating the half vector is done outside of getLightSpecular, and then passed to the function to prevent multiple evaluations.
Fresnel is calculated using the half vector for lights, and can be enabled by setting the calcFresnel flag to true.