Add support for KHR_materials_sheen (#4460)
* Add support for KHR_materials_sheen

Using the Charlie method for sheen and avoiding to use a LUT (might be revisited for performance reasons).
Break down the combineColor into a single chunk with defines. Combine is now a series of adds, and can allow for weighting layers, for example with the new sheen. The old combine functions have been deleted.
Removed the checks of clearcoat > 0 in the lit shader as it's a bool and shouldn't be necessary.

* Fix rebase.

* Added documentation for the sheen parameters.

* Adding sheen parameters to the parameters list and fixup list.

* Fixed error caused by force push.