Add support for KHR_materials_variants
The API lives entirely in the GLB container, allowing the user to get the list of variants and apply it to a mesh.
The variants are parsed, and the data for the variants are stored on the resource level of the container, thus avoiding any intrusion into any engine types.