Reduced passing of backend arguments struct.
Some functions only need one or two arguments, so it would make sense to reuse them now when we don't rely on globals in the backend. One such case is for getLightSpecular, which can just be provided with another gloss and normal for the clearcoat.
Some other functions, like fresnelSchlick, requires compile time conditional arguments to be read, making it and functions that call it bad candidates.
Renamed Frontend struct to LitShaderArgs.
Renamed glossiness to gloss.