Further separation of frontend and backend shaders (#5104)
* Just some spacing

* Make use of frontend struct to pass arguments to backend.

Converted lit-shader to use chunk builders, allowing the backend evaluation to be sectioned into a single function.
Backend chunks read arguments from the frontend exclusively, no longer relying on frontend globals.

* Fix lint

* Moved frontend to shader chunk and made a bunch of smaller fixes.

* Separation of lit shader parameters based on effect.

* Removed d, cc and s prefixes from frontend struct.

Renamed normalW to worldNormal.
Fixed some places where we don't use the frontend struct parameters.

* Fix places where frontend is not being passed down.

* Reduced passing of backend arguments struct.

Some functions only need one or two arguments, so it would make sense to reuse them now when we don't rely on globals in the backend. One such case is for getLightSpecular, which can just be provided with another gloss and normal for the clearcoat.
Some other functions, like fresnelSchlick, requires compile time conditional arguments to be read, making it and functions that call it bad candidates.
Renamed Frontend struct to LitShaderArgs.
Renamed glossiness to gloss.

* Remove test code

* Merge with main sheen specular occlusion.

* Fix occludeSpecular to use raw arguments instead of accepting the LitShaderArgument struct.

* Fix clustered light evaluation using world normal as argument instead of the entire lit arguments struct.

* Fix for LTC lights

* Split combinePS arguments into separate variables

* Removed uvOffset

* Move iridescence, clearcoat and sheen to their own structs.

Helps passing extension parameters such as iridescence as a struct instead of having to unroll all arguments, which makes things a bit tidier.

Removed uvOffset as it's actually used by the frontend.

* Unroll lit shader args for clustered lights

* Fix iridescence

* Fix lightmapdir not being set in litshaderargs.

* Added custom frontend example.

* Revert "Added custom frontend example."

This reverts commit c815e5e0304dbeca8631c303cde5c0d718325ce8.

* Add reflection direction as argument to reflection relevant functions for clarity and reusability.

* Fix typo

* Removed use of reflection direction for sheen.

* Fix lightmap inconsistency.

* Fixed lightmapDir being incorrectly assigned in some conditions.

* Add chunk validation to changes made in backend chunks.

* Fix refraction chunk not passing the reflection vector for cube refractions.

* Move copy over to litShaderArgs at the end of the getter evaluation.

* Added documentation to the litShaderArgs and extension structs.

* Move view direction from global to argument.

* Move dLightPos, dLightDir and dLightDirNorm to args instead of globals.

* Removed backend chunks reliance on dLightPosW, dLightDirW and dLightDirNormW and dShadowCoord.

Instead, these values are passed to their constituent function functions from the point of initialisation.

* Remove reliance on dVertexNormalW.

This variable is already a misnomer. It's not actually the vertex normal, should probably be called geometric normal, as it's the normal extrapolated the vertices, and it's not the surface normal used for lighting.

* Cleanup and removed dependence on dTBN.

* Replace the three attenuation global factors in clustered lighting as clustered lighting would trigger them all to get generated anyways.

* Update chunk validation and group globals in lit shader based on usecase.

* Added back accidentally deleted jsdocs from litshader. Fixed indentation in clusteredlight.js

* Remove iridescenceFresnel from the iridescence arguments as it's supposed to be updated by the backend, and not set by the frontend.

* Fix call to LTC light uv when using clearcoat.

* Rename all the temporary named variables from tX to something more descriptive.

Fixed bug in aniso GGX chunk.