Fixes Adreno 610 and 618 bug.
Moves the lit shader arguments out of the struct and in to globals. The retain the same naming strategy, with litShaderArgs being the first level, and the clearcoat/sheen/iridescence extensions being another layer, so for example litShaderArgs_sheen_gloss gets the sheen gloss value.

Also splits the metalness modulate into two functions, one for albedo and one for specular.