Fix average blocker normalization factor for directional lights
Directional lights can't really normalize the depth difference range because they have no position, so effectively this range becomes gigantic with regards to the distance between the light source and blocker. To fix this, we simply use a defined value to define the normalization value for directional lights, and saturate to avoid oversampling the shadow area.