fix: example thumbnail build (#5522)
* Update index.mjs

The file and content exists in the generated site but the class for this example is commented out.  This bocks thumbnails from being generated.

* Update thumbnails.mjs

Ignore the file so it does not go searching for the definitions (commented or not)

* Update thumbnails.mjs

Ensure temporary filter is noted after building thumbs.

* Update index.mjs

Exclude LitMaterialExample example from build products.