Update post effect queue type mismatch
currently, first argument of addEffect() function is typed as PostEffect which is wrong because it refers to PostEffect class that is declared on the same file, it should have been pointing to pc.PostEffect. but because pc. prefix removed from docs, renamed it to PostEffectEntry, as it is assigned to variable named "newEntry".